Refactor node color logic and move it to a separate file
